requirements.txt
google-generativeai>=0.3.0
rich>=13.7.0
python-dotenv>=1.0.0
typer>=0.9.0 